:building_construction: move login routing functions to a seperate controller
